requirements/base.txt
# =============================================================================
# VansDevBlog Search Service - Base Requirements
# 모든 환경에서 공통으로 필요한 패키지들
# =============================================================================

# Django 기본
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Elasticsearch 연동
django-elasticsearch-dsl==7.3.0
elasticsearch==8.11.0

# API 문서화
drf-yasg==1.21.7

# 데이터베이스 드라이버
PyMySQL==1.1.0              # MariaDB 연결 (Pure Python)
pymongo==4.5.0              # MongoDB 연결

# 환경변수 관리
python-dotenv==1.0.0

# 데이터 검증
pydantic==2.1.1

# 날짜/시간 처리
python-dateutil==2.8.2

# HTTP 요청
requests==2.31.0

